# improved xor performance