Improved xor performance using custom implementation rather than calling `difference` and `union` internally.

## :o: original implementation:
Using `difference` and `union` internally.

|          | xor of 1 set: 10M | xor of 2 sets: 10M & 5M | xor of 3 sets: 10M, 5M, & 3M |
|:---------|:------------------|:------------------------|:-----------------------------|
| test 1:  | 6629.388 ms       | 8918.653 ms             | 11540.040 ms                 |
| test 2:  | 6463.766 ms       | 8907.801 ms             | 11917.361 ms                 |
| test 3:  | 6566.865 ms       | 8812.364 ms             | 11522.295 ms                 |
| test 4:  | 6639.678 ms       | 8727.676 ms             | 11625.841 ms                 |
| test 5:  | 7025.046 ms       | 8848.937 ms             | 11584.064 ms                 |
| average: | 6664.949 ms       | 8843.086 ms             | 11637.920 ms                 |

## :o: xor custom implementation _(option 1)_:
Calls `reusedValues.has` before calling `reusedValues.add`.

|                      | xor of 1 set: 10M | xor of 2 sets: 10M & 5M | xor of 3 sets: 10M, 5M, & 3M |
|:---------------------|:------------------|:------------------------|:-----------------------------|
| test 1:              | 3364.182 ms       | 6053.281 ms             | 7716.744 ms                  |
| test 2:              | 3312.582 ms       | 5920.478 ms             | 7481.409 ms                  |
| test 3:              | 3374.251 ms       | 5827.507 ms             | 7689.054 ms                  |
| test 4:              | 3589.623 ms       | 6137.061 ms             | 7608.236 ms                  |
| test 5:              | 3207.822 ms       | 5797.050 ms             | 7354.333 ms                  |
| average:             | 3369.692 ms       | 5947.075 ms             | 7569.955 ms                  |
| ***relative time:*** | ***50.558%***     | ***67.251%***           | ***65.046%***                |

## :heavy_check_mark: xor custom implementation _(option 2)_:
Always calls `reusedValues.add` instead of checking for a value first.

Notes:
- _relative time_ is a comparison of averages between this implementation and the original.
- On my machine, ~3200ms is how long it takes to copy a Set of size 10M. So this is approximately the minimum threshold we can aim for, given that the function is immutable and returns a copied Set.

|                      | xor of 1 set: 10M | xor of 2 sets: 10M & 5M | xor of 3 sets: 10M, 5M, & 3M |
|:---------------------|:------------------|:------------------------|:-----------------------------|
| test 1:              | 3432.722 ms       | 5572.122 ms             | 6884.741 ms                  |
| test 2:              | 3260.315 ms       | 5673.121 ms             | 6908.560 ms                  |
| test 3:              | 3331.252 ms       | 5913.558 ms             | 6921.396 ms                  |
| test 4:              | 3296.804 ms       | 5689.021 ms             | 7052.191 ms                  |
| test 5:              | 3424.290 ms       | 5797.473 ms             | 7237.603 ms                  |
| average:             | 3349.077 ms       | 5729.059 ms             | 7000.898 ms                  |
| ***relative time:*** | ***50.249%***     | ***64.786%***           | ***60.156%***                |